Update audio.go
fix

+++ audio/audio.go
@@ -161,15 +161,15 @@ func write(buf []byte, size int32) int32 {
 		bufOff = 0
 		bufDiv = 2
 	} else {
-		time.Sleep(time.Millisecond * time.Duration((paPtr-paPlayPtr)/bufThreshold*(bufBlock/int(paRate))))
+		time.Sleep(time.Millisecond * time.Duration((paPtr-paPlayPtr)/bufThreshold*int64(bufBlock/int(paRate))))
 	}
 
 	// time.Sleep(time.Millisecond * time.Duration((paPtr-paPlayPtr)/bufThreshold*(bufBlock/int(paRate))))
 
-	mm := int(min(size/4, bufSize - bufOff))
+	mm := min(int(size/4), int(bufSize - bufOff))
 	for i := 0; i < mm; i++ {
 		p := 4 * (int32(i))
-		paBuf[paPtr-(paPtr/bufSize)*bufSize] = int32(binary.LittleEndian.Uint32(buf[p : p+4])/bufDiv)
+		paBuf[paPtr-(paPtr/bufSize)*bufSize] = int32(binary.LittleEndian.Uint32(buf[p : p+4]))/int32(bufDiv)
 		paPtr++
 		written += 4
 	}
